Move SavedEncounter to  common
