+++ client/Combatant/Combatant.ts
@@ -3,7 +3,7 @@ import * as ko from "knockout";
 import { AbilityScores, StatBlock } from "../../common/StatBlock";
 import { probablyUniqueString } from "../../common/Toolbox";
 import { Encounter } from "../Encounter/Encounter";
-import { SavedCombatant } from "../Encounter/SavedEncounter";
+import { SavedCombatant } from "../../common/SavedEncounter";
 import { Dice } from "../Rules/Rules";
 import { CurrentSettings } from "../Settings/Settings";
 import { Metrics } from "../Utility/Metrics";
