+++ client/Library/EncounterLibrary.ts
@@ -2,7 +2,7 @@ import * as ko from "knockout";
 
 import { ServerListing } from "../../common/Listable";
 import { AccountClient } from "../Account/AccountClient";
-import { SavedCombatant, SavedEncounter } from "../Encounter/SavedEncounter";
+import { SavedCombatant, SavedEncounter } from "../../common/SavedEncounter";
 import { UpdateLegacySavedEncounter } from "../Encounter/UpdateLegacySavedEncounter";
 import { Store } from "../Utility/Store";
 import { Listing, ListingOrigin } from "./Listing";
